Get the transfer Lambda doing proper diags
