+++ terraform/modules/stack/lambda/cloudwatch_log_group.tf
@@ -1,5 +1,7 @@
 resource "aws_cloudwatch_log_group" "cloudwatch_log_group" {
   name = "/aws/lambda/${var.name}"
+
+  retention_in_days = 7
 }
 
 resource "aws_iam_role_policy" "cloudwatch_logs" {
